// store/stories/stories.types.ts
export interface Passage {
	/**
	 * Height of the passage in pixels.
	 */
	height: number;
	/**
	 * Should the passage be drawn highlighted?
	 */
	highlighted: boolean;
	/**
	 * GUID identifying the passage.
	 */
	id: string;
	/**
	 * Left (e.g. X) position of the top-left corner of the passage in pixels.
	 */
	left: number;
	/**
	 * Name of the passage.
	 */
	name: string;
	/**
	 * Is the passage currently selected by the user?
	 */
	selected: boolean;
	/**
	 * ID of the parent story.
	 */
	story: string;
	/**
	 * Passage tags.
	 */
	tags: string[];

	commands: PassageCommand[];
	commandBreakPoint?: CommandBreakPoints;
	/**
	 * Top (e.g. Y) position of the top-left corner of the passage in pixels.
	 */
	top: number;
	/**
	 * Width of the passage in pixels.
	 */
	width: number;
	/**
	 * Random pastel color assigned to this passage.
	 */
	color?: string;
}

export interface Story {
	/**
	 * IFID of the story. An IFID should stay stable when a story is imported or exported.
	 */
	ifid: string;
	/**
	 * GUID identifying the story.
	 */
	id: string;
	/**
	 * When the story was last changed.
	 */
	lastUpdate: Date;
	/**
	 * Name of the story.
	 */
	name: string;
	/**
	 * Passages in the story.
	 */
	passages: Passage[];
	/**
	 * Author-created JavaScript associated with the story.
	 */
	script: string;
	/**
	 * Is the story currently selected by the user?
	 */
	selected: boolean;
	/**
	 * Should passages snap to a grid?
	 */
	snapToGrid: boolean;
	/**
	 * ID of the passage that the story begins at.
	 */
	startPassage: string;
	/**
	 * Name of the story format the story uses.
	 */
	storyFormat: string;
	/**
	 * Version of the story format that this story uses.
	 */
	storyFormatVersion: string;
	/**
	 * Author-created CSS associated with the story.
	 */
	stylesheet: string;
	/**
	 * Tags applied to the story.
	 */
	tags: string[];
	/**
	 * Author-specified colors for passage tags.
	 */
	tagColors: TagColors;
	/**
	 * Zoom level the story is displayed at.
	 */
	zoom: number;

	/**
	 * More config goes here.
	 */
	storySetting: StorySetting,

	/**
	 * Path to the preset file.
	 */
	presetPath?: string;

	/**
	 * Path to the template file (if this is a template story)
	 */
	templatePath?: string;

	/**
	 * Whether remixing is enabled for the story.
	 */
	remixEnabled: boolean;

	/**
	 * Orientation of the story.
	 */
	orientation: 'landscape' | 'portrait';

	/**
	 * Whether the story is archived.
	 */
	archived?: boolean;
}

export type StorySetting = {
	startMenuSetting: any;
	endMenuSetting: any;
	characterDialogSetting: any,
	characterSetting: any,
	choiceMenuSetting: any,
}
// ...
export interface CreatePassageAction {
	type: 'createPassage';
	props: Partial<Passage>;
	storyId: string;
}
// ...
export type TagColors = Record<string, Exclude<Color, 'none'>>;